fix: use named arguments in load_cifar10_data call
`utils.load_cifar10_data` takes in 4 parameters, not 3. The second parameter is `download`. So the previous code is actually wrong by passing in `trans['aug']` where a boolean was expected.
By specifying the argument names, the transform settings will be passed in correctly -- and this is even resilient to changes in the arguments' order.